Move GrpcHeaderNames from internal package to common package
